Fix an issue introduced back in 2.18.0 around more flexible env types
This commit fixes an issue when passing in a mapping which is not an
actual dict as an environment argument. Internally, AsyncSSH uses a
read-only MappingProxyType in places which can trigger this problem
if an env from one session is used when starting another. Thanks go
to Boris Pavlovic for reporting this issue!