When possible, defer calling passphrase callback
When loading encrypted keys, AsyncSSH allows a passphrase to be specified
either directly as a str/bytes valaue or as a callable or coroutine which
takes an argument of the name of the file being loaded, allowing multiple
keys to be loaded with different passphrases in a single operation. However,
until now, the callbacks were always called at the time the keys were loaded.

With this change, the callable/coroutine for a key is only called when the
key is actually used to perform a signing operation. This can be useful
if the callback prompts the end user for the passphrase and you want to
minimize the number of such requests.

This feature only applies when the public key associated with an encrypted
private key is available. This is alwyays the case when using OpenSSH
fotmat, but for older PEM/DER formats, a separate public key might need
to be provided.