Improve handling of KeyboardInterrupt and task cancellation in SCP
This commit makes sure to re-raise CancelledError when an SCP operation
is cancelled, letting the calling code sees the exception and can decide
whether to cancel future operations or not. Thanks go to Viktor Kertesz
for reporting this issue and helping to understand the behavior in
various versions of Python.

This also includes a small change to abort an SCP session when
KeyboardInterrupt is raised, closing the session more quickly.